Split vcfParser in to 2 classes ClinvarFilter and ClinvarMatcher which both implement VcfParser
